+++ application/languages/fr/views.php
@@ -117,6 +117,7 @@ return [
     Episciences_Paper_Logger::CODE_VERSION_REPOSITORY_UPDATED => Episciences_Paper_Logger::$_label[Episciences_Paper_Logger::CODE_VERSION_REPOSITORY_UPDATED],
     Episciences_Paper_Logger::CODE_ACCEPTED_ASK_AUTHORS_FINAL_VERSION => Episciences_Paper_Logger::$_label[Episciences_Paper_Logger::CODE_ACCEPTED_ASK_AUTHORS_FINAL_VERSION],
     Episciences_Paper_Logger::CODE_NEW_REVIEWING_DEADLINE => Episciences_Paper_Logger::$_label[Episciences_Paper_Logger::CODE_NEW_REVIEWING_DEADLINE],
+    Episciences_Paper_Logger::CODE_INBOX_COAR_NOTIFY_REVIEW => Episciences_Paper_Logger::$_label[Episciences_Paper_Logger::CODE_INBOX_COAR_NOTIFY_REVIEW],
 
     // Erreurs OAI
     "The operation ended with an error. Please try again." => "L'opération s'est terminée avec une erreur. Merci de réessayer.",